# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/summary:
    get:
      operationId: getWalletSummary
      tags: [wallet]
      summary: Get wallet balances and earnings
      responses:
        "200":
          description: Wallet summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /wallet/transactions:
    get:
      operationId: listTransactions
      tags: [wallet]
      summary: List wallet transactions
      responses:
        "200":
          description: Wallet transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /referrals/summary:
    get:
      operationId: getReferralSummary
      tags: [wallet]
      summary: Get referral link and performance
      responses:
        "200":
          description: Referral summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /withdrawals:
    get:
      operationId: listWithdrawals
      tags: [wallet]
      summary: List withdrawal requests
      responses:
        "200":
          description: Withdrawal requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Withdrawal"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createWithdrawal
      tags: [wallet]
      summary: Request a wallet withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "201":
          description: Withdrawal request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /deposits:
    get:
      operationId: listDeposits
      tags: [wallet]
      summary: List the signed-in user's deposits
      responses:
        "200":
          description: Deposit records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createDeposit
      tags: [wallet]
      summary: Submit a bank transfer for admin verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Pending deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    WalletSummary:
      type: object
      required: [availableBalance, pendingBalance, totalEarned, totalSpent, referralEarned, withdrawals]
      properties:
        availableBalance: { type: number }
        pendingBalance: { type: number }
        totalEarned: { type: number }
        totalSpent: { type: number }
        referralEarned: { type: number }
        withdrawals: { type: number }
    Transaction:
      type: object
      required: [id, type, description, amount, status, createdAt]
      properties:
        id: { type: string }
        type: { type: string, enum: [task_reward, platform_reward, referral_reward, task_funding, deposit, withdrawal, refund] }
        description: { type: string }
        amount: { type: number }
        status: { type: string, enum: [completed, pending, failed] }
        taskId: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ReferralSummary:
      type: object
      required: [referralCode, referralLink, directReferrals, referralEarnings, recentReferrals]
      properties:
        referralCode: { type: string }
        referralLink: { type: string, format: uri }
        directReferrals: { type: integer }
        referralEarnings: { type: number }
        recentReferrals:
          type: array
          items: { $ref: "#/components/schemas/Referral" }
    Referral:
      type: object
      required: [id, name, joinedAt, earned]
      properties:
        id: { type: string }
        name: { type: string }
        joinedAt: { type: string, format: date-time }
        earned: { type: number }
    Withdrawal:
      type: object
      required: [id, amount, method, status, requestedAt]
      properties:
        id: { type: string }
        amount: { type: number }
        method: { type: string, enum: [bank_transfer, mobile_money, crypto] }
        accountLabel: { type: string }
        status: { type: string, enum: [pending, approved, rejected, paid] }
        note: { type: ["string", "null"] }
        requestedAt: { type: string, format: date-time }
        processedAt: { type: ["string", "null"], format: date-time }
    WithdrawalInput:
      type: object
      required: [amount, method, accountLabel]
      properties:
        amount: { type: number, minimum: 1 }
        method: { type: string, enum: [bank_transfer, mobile_money, crypto] }
        accountLabel: { type: string, minLength: 3, maxLength: 120 }
...
    Deposit:
      type: object
      required: [id, amount, bankName, transferReference, depositReference, transferredAt, status, createdAt]
      properties:
        id: { type: string }
        amount: { type: number }
        bankName: { type: string }
        transferReference: { type: string }
        depositReference: { type: string }
        transferredAt: { type: string, format: date-time }
        proofUrl: { type: ["string", "null"], format: uri }
        status: { type: string, enum: [pending, approved, rejected] }
        note: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        reviewedAt: { type: ["string", "null"], format: date-time }
    DepositInput:
      type: object
      required: [amount, bankName, transferReference, transferredAt]
      properties:
        amount: { type: number, minimum: 1 }
        bankName: { type: string, minLength: 2, maxLength: 120 }
        transferReference: { type: string, minLength: 2, maxLength: 160 }
        transferredAt: { type: string, format: date-time }
        proofUrl: { type: string, format: uri, maxLength: 500 }
...