# lib/api-zod/src/generated/types/depositInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DepositInput {
  /** @minimum 1 */
  amount: number;
  /**
     * @minLength 2
     * @maxLength 120
     */
  bankName: string;
  /**
     * @minLength 2
     * @maxLength 160
     */
  transferReference: string;
  transferredAt: Date;
  /** @maxLength 500 */
  proofUrl?: string;
}